Refactoring packages
