+++ pkg/pgdb/store.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
+	"github.com/pennsieve/pennsieve-go-core/pkg/pgdb/models"
 )
 
 // SQLStore provides the Queries interface and a db instance.
@@ -39,8 +40,8 @@ func (store *SQLStore) execTx(ctx context.Context, fn func(*Queries) error) erro
 }
 
 // ImportFiles creates rows for uploaded files in Packages and Files tables as a transaction
-func (store *SQLStore) ImportFiles(ctx context.Context, records []PackageParams) ([]Package, error) {
-	var result []Package
+func (store *SQLStore) ImportFiles(ctx context.Context, records []models.PackageParams) ([]models.Package, error) {
+	var result []models.Package
 
 	err := store.execTx(ctx, func(q *Queries) error {
 		// TODO: add packages
